working on authorization for routes

+++ app/api/product_routes.py
@@ -16,42 +16,48 @@ def get_products():
 @product_routes.route('/', methods=['POST'])
 def create_product():
     # get seller_id from session
-    print('current_user', current_user)
-    seller_id = current_user
-
+    # seller_id = current_user
     form = ProductForm()
-    form['csrf_token'].data = request.cookies['csrf_token']
-    if form.validate_on_submit():
-        product = Product(
-            name=form.data['name'],
-            description=form.data['description'],
-            category=form.data['category'],
-            price=form.data['price'],
-            quantity=form.data['quantity'],
-            seller_id=seller_id
-        )
-        db.session.add(product)
-        db.session.commit()
-        return product.to_dict()
-    return {'errors': form.errors}, 401
+    if current_user.is_authenticated:
+        user = current_user.to_dict()
+        seller_id = user['id']
+        form['csrf_token'].data = request.cookies['csrf_token']
+        if form.validate_on_submit():
+            product = Product(
+                name=form.data['name'],
+                description=form.data['description'],
+                category=form.data['category'],
+                price=form.data['price'],
+                quantity=form.data['quantity'],
+                seller_id=seller_id
+            )
+            db.session.add(product)
+            db.session.commit()
+            return product.to_dict()
+        return {'errors': form.errors}, 401
+    return {'errors': 'Unauthorized'}, 403
 
 # Update a product
 @product_routes.route('/<int:id>', methods=['PUT'])
 def update_product(id):
-    product = Product.query.get(id)
     form = ProductForm()
-    form['csrf_token'].data = request.cookies['csrf_token']
-    if form.validate_on_submit():
-        product.name = form.data['name']
-        product.description = form.data['description']
-        product.category = form.data['category']
-        product.price = form.data['price']
-        product.quantity = form.data['quantity']
-        product.images = form.data['images']
-        product.seller_id = form.data['seller_id']
-        db.session.commit()
-        return product.to_dict()
-    return {'errors': form.errors}, 401
+    if current_user.is_authenticated:
+        user = current_user.to_dict()
+        product = Product.query.get(id)
+        
+        form['csrf_token'].data = request.cookies['csrf_token']
+        if form.validate_on_submit():
+            product.name = form.data['name']
+            product.description = form.data['description']
+            product.category = form.data['category']
+            product.price = form.data['price']
+            product.quantity = form.data['quantity']
+            product.images = form.data['images']
+            product.seller_id = form.data['seller_id']
+            db.session.commit()
+            return product.to_dict()
+        return {'errors': form.errors}, 401
+    return {'errors': 'Unauthorized'}, 403
 
 # Get a single product
 @product_routes.route('/<int:id>')
@@ -62,8 +68,12 @@ def get_product(id):
 # Delete a product
 @product_routes.route('/<int:id>', methods=['DELETE'])
 def delete_product(id):
-    product = Product.query.get(id)
-    db.session.delete(product)
-    db.session.commit()
-    return {'message': 'Product deleted'}
+    if current_user.is_authenticated:
+        user = current_user.to_dict()
+        product = Product.query.get(id)
+        db.session.delete(product)
+        db.session.commit()
+        return {'message': 'Product deleted'}
+    return {'errors': 'Unauthorized'}, 403
+
